refactor use common package exceptions
