+++ src/pururu/domain/poll_system/poll_resolution_factory.py
@@ -1,4 +1,4 @@
-from pururu.domain.exceptions import PururuException
+from pururu.common.exceptions import PururuException
 from pururu.domain.entities import PollResolutionType
 from pururu.domain.poll_system.poll_resolution_strategy import SendMessagePollResolution, PollResolutionStrategy
