+++ src/pururu/infrastructure/adapters/discord/discord_service_adapter.py
@@ -4,7 +4,7 @@ import discord
 
 from pururu.common import utils
 from pururu.domain.entities import Message, Poll
-from pururu.domain.exceptions import DiscordServiceException
+from pururu.common.exceptions import DiscordServiceException
 from pururu.domain.services.discord_service import DiscordInterface
 from pururu.infrastructure.adapters.discord.discord_bot import PururuDiscordBot
