+++ src/tests/test_domain/test_poll_system/test_poll_resolution_factory.py
@@ -3,7 +3,7 @@ from unittest.mock import Mock
 import pytest
 from hamcrest import assert_that, instance_of, equal_to
 
-from pururu.domain.exceptions import PururuException
+from pururu.common.exceptions import PururuException
 from pururu.domain.entities import PollResolutionType
 from pururu.domain.poll_system.poll_resolution_factory import PollResolutionFactory
 from pururu.domain.poll_system.poll_resolution_strategy import SendMessagePollResolution
